+++ test/tools/create-configs.ts
@@ -1,0 +1,7 @@
+import { BundlibPkgJson, configsFromPkg } from "../../src/api";
+
+async function createConfigs(cwd: string, dev: boolean, pkgJson?: BundlibPkgJson) {
+  return pkgJson ? configsFromPkg(cwd, dev, pkgJson) : configsFromPkg(cwd, dev);
+}
+
+export default createConfigs;